+++ src/frontend/stacked_windows/settings/dependencies.py
@@ -50,11 +50,7 @@ class DependencySettings(BaseSettings):
         self.add_layout(self._build_dependency_layout(*self.ffmpeg_widgets))
         self.add_layout(self._build_dependency_layout(*self.frame_forge_widgets))
         self.add_layout(self._build_dependency_layout(*self.mkbrr_widgets))
-        self.add_layout(self.reset_layout)
-
-        # shift the spacer under neath the widgets
-        self.inner_layout.removeItem(self._spacer_item)
-        self.inner_layout.addSpacerItem(self._spacer_item)
+        self.add_layout(self.reset_layout, add_stretch=True)
 
         self._load_saved_settings()